- changing all double variables to int

+++ java_dishonest_writers/src/main/java/com/caykeprudente/ClientHandler.java
@@ -76,9 +76,9 @@ public class ClientHandler implements Runnable {
         else
             request.put( "type", Define.write_back);
 
-        request.put( "timestamp", data.timestamp.intValue());
+        request.put( "timestamp", data.timestamp);
         request.put( "variable", data.value);
-        request.put( "request_code", data.request_code.intValue());
+        request.put( "request_code", data.request_code);
         request.put( "echoes", echoesArray);
         injectClientInfo(request);
 
@@ -115,7 +115,7 @@ public class ClientHandler implements Runnable {
             client.lock.lock();
             client.increment_timestamp_by = 1;
             client.lock.unlock();
-            
+
             if (this.client.verbose > 0)
                 System.out.println("Tried to write, but there is a newest data already");
         }
@@ -134,7 +134,7 @@ public class ClientHandler implements Runnable {
     private void readFromServer() {
         Map<String, Object> request = new HashMap<String, Object>();
         request.put( "type", Define.read);
-        request.put( "request_code", data.request_code.intValue());
+        request.put( "request_code", data.request_code);
         injectClientInfo(request);
 
         Socket clientSocket = null;
@@ -197,7 +197,7 @@ public class ClientHandler implements Runnable {
     private void readTimestampFromServer() {
         Map<String, Object> request = new HashMap<String, Object>();
         request.put( "type", Define.read_timestamp);
-        request.put( "request_code", data.request_code.intValue());
+        request.put( "request_code", data.request_code);
         injectClientInfo(request);
 
         Socket clientSocket = null;
@@ -261,9 +261,9 @@ public class ClientHandler implements Runnable {
     private void readEchoeFromServer() {
         Map<String, Object> request = new HashMap<String, Object>();
         request.put( "type", Define.get_echoe);
-        request.put( "request_code", data.request_code.intValue());
+        request.put( "request_code", data.request_code);
         request.put( "variable", data.value);
-        request.put( "timestamp", data.timestamp.intValue());
+        request.put( "timestamp", data.timestamp);
         injectClientInfo(request);
 
         Socket clientSocket = null;
@@ -294,10 +294,10 @@ public class ClientHandler implements Runnable {
 
             client.lock.lock();
             if (client.echoes.size() < client.quorum-1) {
-                client.echoes.add(new Pair<Integer, String>(responseData.server_id.intValue(), responseData.data_signature));
+                client.echoes.add(new Pair<Integer, String>(responseData.server_id, responseData.data_signature));
             }
             else if (client.echoes.size() == client.quorum-1) {
-                client.echoes.add(new Pair<Integer, String>(responseData.server_id.intValue(), responseData.data_signature));
+                client.echoes.add(new Pair<Integer, String>(responseData.server_id, responseData.data_signature));
                 client.semaphore.release();
             }
             else {
@@ -339,7 +339,7 @@ public class ClientHandler implements Runnable {
     }
 
     private void injectClientInfo(Map<String, Object> request) {
-        request.put(Define.client_id, this.client.id.intValue());
+        request.put(Define.client_id, this.client.id);
         request.put(Define.server_plataform, Define.plataform);
     }
 }
