+++ java_dishonest_writers/src/main/java/com/caykeprudente/ResponseData.java
@@ -21,15 +21,15 @@ public class ResponseData {
     */
 
     String value;
-    Double timestamp;
+    int timestamp;
     String data_signature;
-    Double server_id;
+    int server_id;
     List<Pair<Integer, String>> echoes;
     Pair<String, Integer> server;
 
-    Double request_code;
+    int request_code;
 
-    public ResponseData(String value, Double timestamp, String data_signature, Double server_id, Double request_code, Pair<String, Integer> server, List<Pair<Integer, String>> echoes) {
+    public ResponseData(String value, int timestamp, String data_signature, int server_id, int request_code, Pair<String, Integer> server, List<Pair<Integer, String>> echoes) {
         this.value = value;
         this.timestamp = timestamp;
         this.data_signature = data_signature;
@@ -41,11 +41,11 @@ public class ResponseData {
 
     public ResponseData(LinkedTreeMap<String, Object> dictionary, Pair<String, Integer> server) {
         LinkedTreeMap<String, Object> dataDict = (LinkedTreeMap<String, Object>) dictionary.get(Define.data);
-        timestamp = (Double) dataDict.get(Define.timestamp);
+        timestamp = ((Double) dataDict.get(Define.timestamp)).intValue();
         value = (String) dataDict.get(Define.variable);
         data_signature = (String) dataDict.get(Define.data_signature);
 
-        server_id = (Double) dictionary.get(Define.server_id);
+        server_id = ((Double) dictionary.get(Define.server_id)).intValue();
         this.server = server;
     }
 }
